Implemented get_success_url() for ContactCreateView
The former implementation using post did not work as the data is not
live as at the time I am requesting for object.pk.
With get_success_url, it seems POST concludes - with the object created,
before it calls get_success_url.
Returning to the form after creation, instead of the contacts-list is
better experience.